requirements.txt
Django==2.2
django-filter==2.1.0
python-monkey-business==1.0.0
pytz==2018.9
six==1.12.0
sqlparse==0.3.0
